Revert to commit prior to wrapper separation
- The repo has been refactored such that it now uses the boston dynamics
API wrapper that the BDAII developed (which is a good thing). However,
this means that to add the ability to send trajectories to the arm
instead of single points, we'd have to make a PR to that repo instead of
the heuristicus/spot_ros repo. We're just hacking in changes for now to
get a demo working so we're doing this the cludgy way.
- Ultimately, we'll want to pull in the spot wrapper refactor (likely in
the summer).